backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
yfinance==0.2.18
requests==2.31.0
python-dotenv==1.0.0